DRY up amount form input normalization
