+++ src/features/wallet/send/SendView.tsx
@@ -22,6 +22,7 @@ import {
   makeBuyerTransferHotspotTxn,
   makePaymentTxn,
   makeSellerTransferHotspotTxn,
+  formatAmountInput,
 } from '../../../utils/transactions'
 import {
   getAccount,
@@ -39,7 +40,7 @@ import {
 import { getAddress } from '../../../utils/secureAccount'
 import Text from '../../../components/Text'
 import useSubmitTxn from '../../../hooks/useSubmitTxn'
-import { decimalSeparator, groupSeparator, locale } from '../../../utils/i18n'
+import { decimalSeparator, groupSeparator } from '../../../utils/i18n'
 import { useAppDispatch } from '../../../store/store'
 import {
   fetchCurrentOraclePrice,
@@ -157,7 +158,7 @@ const SendView = ({ scanResult, sendType, hotspot, isSeller }: Props) => {
             parseFloat(scanAmount),
             CurrencyType.networkToken,
           )
-          amount = formatAmount(String(scanAmount))
+          amount = formatAmountInput(String(scanAmount))
         }
         return {
           id: `transfer${i}`,
@@ -228,24 +229,6 @@ const SendView = ({ scanResult, sendType, hotspot, isSeller }: Props) => {
     transferData?.amountToSeller,
     hasValidActivity,
   ])
-
-  const formatAmount = (formAmount: string) => {
-    if (formAmount === decimalSeparator || formAmount.includes('NaN')) {
-      return `0${decimalSeparator}`
-    }
-    const rawInteger = (formAmount.split(decimalSeparator)[0] || formAmount)
-      .split(groupSeparator)
-      .join('')
-    const integer = parseInt(rawInteger, 10).toLocaleString(locale)
-    let decimal = formAmount.split(decimalSeparator)[1]
-    if (integer === 'NaN') {
-      return ''
-    }
-    if (decimal && decimal.length >= 9) decimal = decimal.slice(0, 8)
-    return formAmount.includes(decimalSeparator)
-      ? `${integer}${decimalSeparator}${decimal}`
-      : integer
-  }
 
   const getNonce = (): number => {
     if (!account?.speculativeNonce) return 1
