+++ src/utils/transactions.ts
@@ -142,3 +142,21 @@ export const isPayer = (
 
 export const isPendingTransaction = (item: unknown) =>
   !!(item as PendingTransaction).createdAt
+
+export const formatAmountInput = (formAmount: string) => {
+  if (formAmount === decimalSeparator || formAmount.includes('NaN')) {
+    return `0${decimalSeparator}`
+  }
+  const rawInteger = (formAmount.split(decimalSeparator)[0] || formAmount)
+    .split(groupSeparator)
+    .join('')
+  const integer = parseInt(rawInteger, 10).toLocaleString(locale)
+  let decimal = formAmount.split(decimalSeparator)[1]
+  if (integer === 'NaN') {
+    return ''
+  }
+  if (decimal && decimal.length >= 9) decimal = decimal.slice(0, 8)
+  return formAmount.includes(decimalSeparator)
+    ? `${integer}${decimalSeparator}${decimal}`
+    : integer
+}
